firewall_toggle/requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
pyOpenSSL==22.1.0
python-dotenv==0.21.0
requests==2.28.1
unificontrol==0.2.9
urllib3==1.26.13
Werkzeug==2.2.2
